Implemented some methods in MovieService/Controller. Changed Review/Role to hold a whole instance of the foreign key(User User instead of int UserId)

+++ Squids-Movies-App/SquidsMovieApp.Logic/Contracts/IMovieService.cs
@@ -16,7 +16,7 @@ namespace SquidsMovieApp.Logic.Contracts
         IEnumerable<ParticipantModel> GetAllParticipantsPerMovie(MovieModel movie);
         void AddMovieParticipant(MovieModel movie, ParticipantModel participant,
             string roleName);
-        float GetRating(MovieModel movie);
+        double GetRating(MovieModel movie);
         //IEnumerable<ReviewModel> GetReviews(MovieModel movie);
         IEnumerable<ParticipantModel> GetActors(MovieModel movie);
         IEnumerable<ParticipantModel> GetDirectors(MovieModel movie);
